Moving config provider to separate folder.
